Updated Signature Query
- Existing query searching by project id (CNCF project with 1000's of
signatures) and then sub-filters by other criteria resulting in the code
needing to iterate over potentially thousands of records.  This resulted
in unexpected errors in the python pynamodb library when retrieving the
results. This patch is to add a specific signature query to search
by company_id then filter by project which, in one case, reduces the
number of records returned from 1000's to less than 10.  As a result the
query is much faster and avoids the volume/generator/pagination issue we
are experiencing.

Resolves: #486

Signed-off-by: David Deal <dealako@gmail.com>