[#1037] Add CCLA Filter to Approve List Check
- Added additional filters to the signature check for users where we do
not have a user record or company affiliation. The prior logic did not
filter on CCLAs or approved/signed flags and therefore was querying
nearly all the gRPC records in the database and painstakingly going
through each record looking for 'approved lists' to compare. This is to
not approve the user, but to give them an indication that they are on an
approved list, but they have not affiliated with a company.  The
new/updated logic leverages the correct filter flags such that only CCLA
signatures will be inspected.

Signed-off-by: David Deal <dealako@gmail.com>