Resolved Corporate Console Lib Problems
- Resolved lodash: CVE-2020-8203 low severity Vulnerable versions: <
4.17.19 Patched version: 4.17.19 Versions of lodash prior to 4.17.19 are
vulnerable to Prototype Pollution. The function zipObjectDeep allows a
malicious user to modify the prototype of Object if the property
identifiers are user-supplied. Being affected by this issue requires
zipping objects based on user-provided property arrays.
- Resolved kind-of: CVE-2019-20149 low severity Vulnerable versions: >=
6.0.0, < 6.0.3 Patched version: 6.0.3 Versions of kind-of 6.x prior to
6.0.3 are vulnerable to a Validation Bypass. A maliciously crafted
object can alter the result of the type check, allowing attackers to
bypass the type checking validation.
- Resolved minimist: CVE-2020-7598 low severity Vulnerable versions: >=
1.0.0, < 1.2.3 Patched version: 1.2.3 Affected versions of minimist are
vulnerable to prototype pollution. Arguments are not properly sanitized,
allowing an attacker to modify the prototype of Object, causing the
addition or modification of an existing property that will exist on all
objects.  Parsing the argument --__proto__.y=Polluted adds a y property
with value Polluted to all objects. The argument --__proto__=Polluted
raises and uncaught error and crashes the application.  This is
exploitable if attackers have control over the arguments being passed to
minimist.

Signed-off-by: David Deal <dealako@gmail.com>