[#2337] Resolved Unenroll Issue
- Discovered and resolved issue when enrolling then unenrolling a
project. Problem was that when the Project Group is deleted followed by
a repo disable, the repo disable triggers a recount of the number of
repos for a given CLA Group. As a result of this activity, the Project
CLA Group table is updated as well. Resolved issue by trapping for
missing table entry.

Signed-off-by: David Deal <dealako@gmail.com>