Added Sort/Most Recent to Preview
- While testing update CLA Group template, I noticed that we were not
returning the most recent CLA template. Updated to look at major, minor,
and created date fields.
- For Projects CLA Group API - added a service layer - not sure why we
didn't have a proper service layer in the first place, that was weird.
Also, added the ctx variable to the projects cla group repository APIs.
This impacted a number of files in order to add that silly variable to
make things consistent.  It's done now.

Signed-off-by: David Deal <dealako@gmail.com>